Added selection caret restoration hook
