+++ packages/text-annotator-react/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useRestoreSelectionCaret } from './useRestoreSelectionCaret';
