+++ packages/text-annotator-react/src/index.ts
@@ -1,4 +1,5 @@
 export * from './tei';
+export * from './hooks';
 export * from './TextAnnotator';
 export * from './TextAnnotatorPopup';
 export * from './TextAnnotatorPlugin';
